+++ src/modules/computePlans/ComputePlansSlice.ts
@@ -3,7 +3,6 @@ import { ComputePlanT } from './ComputePlansTypes';
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 
 import { PaginatedApiResponse } from '@/modules/common/CommonTypes';
-import TasksApi from '@/modules/tasks/TasksApi';
 import {
     AggregatetupleT,
     CompositeTraintupleT,
@@ -103,7 +102,7 @@ export const retrieveComputePlan = createAsyncThunk<
 
 export interface retrieveComputePlanTasksArgs {
     computePlanKey: string;
-    page?: number;
+    page: number;
 }
 
 export const retrieveComputePlanTrainTasks = createAsyncThunk<
@@ -112,23 +111,10 @@ export const retrieveComputePlanTrainTasks = createAsyncThunk<
     { rejectValue: string }
 >('computePlans/getTrainTasks', async ({ computePlanKey, page }, thunkAPI) => {
     try {
-        let response;
-
-        if (page === undefined) {
-            response = await TasksApi.listTraintuples([
-                {
-                    asset: 'traintuple',
-                    key: 'compute_plan_key',
-                    value: computePlanKey,
-                },
-            ]);
-        } else {
-            response = await ComputePlansApi.listComputePlanTraintuples(
-                computePlanKey,
-                page
-            );
-        }
-
+        const response = await ComputePlansApi.listComputePlanTraintuples(
+            computePlanKey,
+            page
+        );
         return response.data;
     } catch (err) {
         return thunkAPI.rejectWithValue(err.response.data);
@@ -141,23 +127,10 @@ export const retrieveComputePlanTestTasks = createAsyncThunk<
     { rejectValue: string }
 >('computePlans/getTestTasks', async ({ computePlanKey, page }, thunkAPI) => {
     try {
-        let response;
-
-        if (page === undefined) {
-            response = await TasksApi.listTesttuples([
-                {
-                    asset: 'testtuple',
-                    key: 'compute_plan_key',
-                    value: computePlanKey,
-                },
-            ]);
-        } else {
-            response = await ComputePlansApi.listComputePlanTesttuples(
-                computePlanKey,
-                page
-            );
-        }
-
+        const response = await ComputePlansApi.listComputePlanTesttuples(
+            computePlanKey,
+            page
+        );
         return response.data;
     } catch (err) {
         return thunkAPI.rejectWithValue(err.response.data);
@@ -172,23 +145,10 @@ export const retrieveComputePlanAggregateTasks = createAsyncThunk<
     'computePlans/getAggregateTasks',
     async ({ computePlanKey, page }, thunkAPI) => {
         try {
-            let response;
-
-            if (page === undefined) {
-                response = await TasksApi.listAggregatetuples([
-                    {
-                        asset: 'aggregatetuple',
-                        key: 'compute_plan_key',
-                        value: computePlanKey,
-                    },
-                ]);
-            } else {
-                response = await ComputePlansApi.listComputePlanAggregatetuples(
-                    computePlanKey,
-                    page
-                );
-            }
-
+            const response = await ComputePlansApi.listComputePlanAggregatetuples(
+                computePlanKey,
+                page
+            );
             return response.data;
         } catch (err) {
             return thunkAPI.rejectWithValue(err.response.data);
@@ -204,23 +164,10 @@ export const retrieveComputePlanCompositeTasks = createAsyncThunk<
     'computePlans/getCompositeTasks',
     async ({ computePlanKey, page }, thunkAPI) => {
         try {
-            let response;
-
-            if (page === undefined) {
-                response = await TasksApi.listCompositeTraintuples([
-                    {
-                        asset: 'composite_traintuple',
-                        key: 'compute_plan_key',
-                        value: computePlanKey,
-                    },
-                ]);
-            } else {
-                response = await ComputePlansApi.listComputePlanCompositeTraintuples(
-                    computePlanKey,
-                    page
-                );
-            }
-
+            const response = await ComputePlansApi.listComputePlanCompositeTraintuples(
+                computePlanKey,
+                page
+            );
             return response.data;
         } catch (err) {
             return thunkAPI.rejectWithValue(err.response.data);